fix: support deserializing `include_granted_scopes` from JS clients
Refactors type layout for improved maintainability.
