+++ src/client/types/authorization.rs
@@ -1,168 +1,10 @@
-use chrono::Utc;
-use oauth2::{
-    AuthorizationCode, CsrfToken, ExtraTokenFields, PkceCodeVerifier, RefreshToken, TokenResponse,
-};
-use serde::{Deserialize, Deserializer, Serialize};
-use std::{borrow::Cow, fmt::Display, ops::Deref, str::FromStr};
+use std::{borrow::Cow, fmt, str::FromStr};
+
+use oauth2::{CsrfToken, PkceCodeVerifier};
+use serde::{Deserialize, Serialize, de};
 use url::Url;
 
-use super::GoogleOAuthTokenResponse;
-
-/// Extra fields returned in Google's token response. We expect the `id_token`
-/// to be present so we can validate and extract user identity from it.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct GoogleOAuthExtraTokenFields {
-    id_token: String,
-}
-
-impl ExtraTokenFields for GoogleOAuthExtraTokenFields {}
-
-impl GoogleOAuthExtraTokenFields {
-    pub fn id_token(&self) -> &str {
-        &self.id_token
-    }
-}
-
-/// Wraps a successful OAuth2 token exchange and provides a redirect URL
-/// to send the user to after login.
-#[derive(Debug, Clone)]
-pub struct AuthSuccessRedirectResponse {
-    pub redirect_url: Url,
-    pub token: GoogleOAuthTokenResponse,
-}
-
-impl AuthSuccessRedirectResponse {
-    pub fn new(redirect_url: Url, token: GoogleOAuthTokenResponse) -> Self {
-        Self {
-            redirect_url,
-            token,
-        }
-    }
-}
-
-impl Deref for AuthSuccessRedirectResponse {
-    type Target = GoogleOAuthTokenResponse;
-    fn deref(&self) -> &Self::Target {
-        &self.token
-    }
-}
-
-impl Display for AuthSuccessRedirectResponse {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        let issued_at = Utc::now().timestamp();
-        let expires_in = self.expires_in().map(|d| d.as_secs()).unwrap_or(0);
-        write!(
-            f,
-            "{}#access_token={}&id_token={}&expires_in={}&issued_at={}",
-            self.redirect_url,
-            self.access_token().secret(),
-            self.extra_fields().id_token(),
-            expires_in,
-            issued_at
-        )
-    }
-}
-
-/// Wraps a successful OAuth2 token exchange and provides a redirect URL
-/// to send the user to after login.
-#[derive(Debug, Clone)]
-pub struct AuthErrorRedirectResponse {
-    pub redirect_url: Url,
-    pub error: String,
-}
-
-impl AuthErrorRedirectResponse {
-    pub fn new(redirect_url: Url, error: impl AsRef<str>) -> Self {
-        Self {
-            redirect_url,
-            error: error.as_ref().to_string(),
-        }
-    }
-}
-
-impl Display for AuthErrorRedirectResponse {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(
-            f,
-            "{}#error={}",
-            self.redirect_url,
-            urlencoding::encode(&self.error)
-        )
-    }
-}
-
-#[derive(Debug)]
-pub struct ExchangeAuthorizationCodeRequest {
-    pub(crate) code: AuthorizationCode,
-    pub(crate) pkce_verifier: PkceCodeVerifier,
-}
-
-#[derive(Debug, Clone, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct ExchangeRefreshTokenRequest {
-    pub(crate) refresh_token: RefreshToken,
-}
-
-impl ExchangeRefreshTokenRequest {
-    pub fn new(refresh_token: impl AsRef<str>) -> Self {
-        Self {
-            refresh_token: RefreshToken::new(refresh_token.as_ref().to_owned()),
-        }
-    }
-}
-
-#[derive(Debug, Clone, Serialize)]
-#[serde(rename_all = "camelCase")]
-pub struct ExchangeRefreshTokenResponse {
-    pub(crate) access_token: String,
-    pub(crate) id_token: String,
-    pub(crate) issued_at: i64,
-    pub(crate) expires_in: u64,
-}
-
-impl From<GoogleOAuthTokenResponse> for ExchangeRefreshTokenResponse {
-    fn from(value: GoogleOAuthTokenResponse) -> Self {
-        let issued_at = Utc::now().timestamp();
-        let expires_in = value.expires_in().map(|d| d.as_secs()).unwrap_or(0);
-        let access_token = value.access_token().clone();
-        Self {
-            access_token: access_token.into_secret(),
-            id_token: value.extra_fields().id_token().to_owned(),
-            issued_at,
-            expires_in,
-        }
-    }
-}
-
-#[derive(Debug, Clone, Copy, Serialize, Deserialize)]
-pub struct ExtraParam(&'static str);
-
-impl Deref for ExtraParam {
-    type Target = str;
-    fn deref(&self) -> &Self::Target {
-        self.0
-    }
-}
-
-impl ExtraParam {
-    pub const INCLUDE_GRANTED_SCOPES: ExtraParam = ExtraParam("include_granted_scopes");
-    pub const LOGIN_HINT: ExtraParam = ExtraParam("login_hint");
-    pub const ACCESS_TYPE: ExtraParam = ExtraParam("access_type");
-    pub const PROMPT: ExtraParam = ExtraParam("prompt");
-
-    pub fn into_cow<'a>(&self) -> Cow<'a, str> {
-        Cow::Borrowed(self.0)
-    }
-}
-
-pub trait IntoExtraParam<'a> {
-    fn into_extra_param(self) -> (ExtraParam, Cow<'a, str>);
-}
-
-pub trait ToExtraParams<'a> {
-    /// Converts **self** into a list of query parameters for the authorization request.
-    fn to_extra_params(&self) -> Vec<(ExtraParam, Cow<'a, str>)>;
-}
+use super::extra_params::{ExtraParam, IntoExtraParam, ToExtraParams};
 
 /// Indicates whether your application can refresh access tokens
 /// when the user is not present at the browser.
@@ -183,7 +25,7 @@ pub enum AccessType {
     Offline,
 }
 
-impl Display for AccessType {
+impl fmt::Display for AccessType {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
         match self {
             Self::Offline => write!(f, "offline"),
@@ -234,7 +76,7 @@ impl FromStr for Prompt {
     }
 }
 
-impl Display for Prompt {
+impl fmt::Display for Prompt {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
         match self {
             Self::None => write!(f, "none"),
@@ -264,7 +106,10 @@ pub struct RequestAccessTokenExtraParams {
     /// is granted, the authorization will include any previous authorizations granted
     /// to this user/application combination for other scopes; see Incremental authorization.
     /// Note that you cannot do incremental authorization with the Installed App flow.
-    #[serde(default = "return_granted_scopes")]
+    #[serde(
+        default = "include_granted_scopes",
+        deserialize_with = "deserialize_bool"
+    )]
     pub(crate) include_granted_scopes: bool,
 
     /// If your application knows which user is trying to authenticate, it can use
@@ -285,35 +130,7 @@ pub struct RequestAccessTokenExtraParams {
     pub(crate) prompt: Vec<Prompt>,
 }
 
-#[derive(Deserialize)]
-#[serde(untagged)]
-enum PromptInput {
-    List(Vec<Prompt>),
-    String(Prompt),
-}
-
-fn deserialize_prompt_vec<'de, D>(deserializer: D) -> Result<Vec<Prompt>, D::Error>
-where
-    D: Deserializer<'de>,
-{
-    let input = PromptInput::deserialize(deserializer)?;
-
-    match input {
-        PromptInput::List(list) => Ok(list),
-        PromptInput::String(s) => s
-            .to_string()
-            .split(',')
-            .filter(|part| !part.trim().is_empty())
-            .map(|part| {
-                Prompt::from_str(part.trim()).map_err(|_| {
-                    serde::de::Error::custom(format!("Invalid prompt value: '{}'", part.trim()))
-                })
-            })
-            .collect(),
-    }
-}
-
-fn return_granted_scopes() -> bool {
+fn include_granted_scopes() -> bool {
     true
 }
 
@@ -345,6 +162,68 @@ impl<'a> ToExtraParams<'a> for RequestAccessTokenExtraParams {
         params.push(self.prompt.clone().into_extra_param());
 
         params
+    }
+}
+
+fn deserialize_bool<'de, D>(deserializer: D) -> Result<bool, D::Error>
+where
+    D: de::Deserializer<'de>,
+{
+    struct StringOrBoolVisitor;
+
+    impl<'de> de::Visitor<'de> for StringOrBoolVisitor {
+        type Value = bool;
+
+        fn expecting(&self, formatter: &mut std::fmt::Formatter) -> std::fmt::Result {
+            formatter.write_str(r#""true", "false", true, or false"#)
+        }
+
+        fn visit_bool<E>(self, v: bool) -> Result<Self::Value, E> {
+            Ok(v)
+        }
+
+        fn visit_str<E>(self, v: &str) -> Result<Self::Value, E>
+        where
+            E: de::Error,
+        {
+            match v {
+                "true" => Ok(true),
+                "false" => Ok(false),
+                _ => Err(de::Error::unknown_variant(v, &["true", "false"])),
+            }
+        }
+    }
+
+    deserializer.deserialize_any(StringOrBoolVisitor)
+}
+
+pub(in crate::client) fn deserialize_prompt_vec<'de, D>(
+    deserializer: D,
+) -> Result<Vec<Prompt>, D::Error>
+where
+    D: de::Deserializer<'de>,
+{
+    #[derive(Deserialize)]
+    #[serde(untagged)]
+    enum PromptInput {
+        List(Vec<Prompt>),
+        String(Prompt),
+    }
+
+    let input = PromptInput::deserialize(deserializer)?;
+
+    match input {
+        PromptInput::List(list) => Ok(list),
+        PromptInput::String(s) => s
+            .to_string()
+            .split(',')
+            .filter(|part| !part.trim().is_empty())
+            .map(|part| {
+                Prompt::from_str(part.trim()).map_err(|_| {
+                    serde::de::Error::custom(format!("Invalid prompt value: '{}'", part.trim()))
+                })
+            })
+            .collect(),
     }
 }
